Refactor Job
