+++ src/duqtools/status.py
@@ -1,7 +1,7 @@
 import logging
-from pathlib import Path
 from time import sleep
 
+from ._job import Job
 from .config import cfg
 from .jetto import JettoSettingsManager
 from .models import WorkDirectory
@@ -11,59 +11,6 @@ info, debug = logger.info, logger.debug
 stream = logging.StreamHandler()
 stream.setFormatter(logging.Formatter('%(message)s'))
 logger.addHandler(stream)
-
-
-class Job:
-
-    def __init__(self, dir: Path):
-        self.dir = Path(dir)
-
-    def __repr__(self):
-        return f'{self.__class__.__name__}({self.dir!r})'
-
-    @property
-    def has_submit_script(self) -> bool:
-        return (self.dir / cfg.submit.submit_script_name).exists()
-
-    @property
-    def has_status(self) -> bool:
-        return (self.dir / cfg.status.status_file).exists()
-
-    @property
-    def is_submitted(self) -> bool:
-        return (self.dir / 'duqtools.submit.lock').exists()
-
-    def status_file_contains(self, msg) -> bool:
-        sf = self.status_file
-        with open(sf, 'r') as f:
-            content = f.read()
-            debug('Checking if content of %s file: %s contains %s', sf,
-                  content, msg)
-            return msg in content
-
-    @property
-    def is_completed(self) -> bool:
-        return self.status_file_contains(cfg.status.msg_completed)
-
-    @property
-    def is_failed(self) -> bool:
-        return self.status_file_contains(cfg.status.msg_failed)
-
-    @property
-    def is_running(self) -> bool:
-        return self.status_file_contains(cfg.status.msg_running)
-
-    @property
-    def in_file(self) -> Path:
-        return self.dir / cfg.status.in_file
-
-    @property
-    def out_file(self) -> Path:
-        return self.dir / cfg.status.out_file
-
-    @property
-    def status_file(self) -> Path:
-        return self.dir / cfg.status.status_file
 
 
 class Status():
